+++ libft/srcs/others/ft_is_positive_int.c
@@ -6,7 +6,7 @@
 /*   By: mdavid <mdavid@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/08 18:35:39 by mdavid            #+#    #+#             */
-/*   Updated: 2020/08/05 10:19:32 by mdavid           ###   ########.fr       */
+/*   Updated: 2020/08/21 13:27:44 by mdavid           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,5 +40,5 @@ int		ft_is_positive_int(char *str)
 		if (result > INTEGER_MAX)
 			return (-1);
 	}
-	return (str[i] == '\0' ? result : -1);
+	return (str[i] == '\0' ? (int)result : -1);
 }
